requirements.txt
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.2.0
Django==3.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
idna==3.4
kavenegar==1.1.2
psycopg2==2.9.6
psycopg2-binary==2.9.6
PyJWT==2.8.0
pytz==2023.3
requests==2.31.0
sqlparse==0.4.4
urllib3==2.0.4
